+++ assets/javascripts/kitten/components/tooltips/tooltip.js
@@ -1,22 +1,22 @@
 import React from 'react'
 import ReactTooltip from 'react-tooltip'
-import ButtonTooltipIcon from 'kitten/components/buttons/button-tooltip-icon'
+import { ButtonQuestionMarkIcon } from 'kitten/components/buttons/button-question-mark-icon'
 
 export default class Tooltip extends React.Component {
   render() {
     const { place,
             children,
             id,
-            ...buttonTooltipIconProps } = this.props
+            ...buttonQuestionMarkIconProps } = this.props
 
     return (
       <div className="k-Tooltip">
-        <ButtonTooltipIcon data-tip
-                           data-for={ id }
-                           data-event="click"
-                           data-dismiss={ id }
-                           aria-describedby={ id }
-                           { ...buttonTooltipIconProps } />
+        <ButtonQuestionMarkIcon data-tip
+                                data-for={ id }
+                                data-event="click"
+                                data-dismiss={ id }
+                                aria-describedby={ id }
+                                { ...buttonQuestionMarkIconProps } />
         <ReactTooltip id={ id }
                       // This is not a mistake, this attribute is called
                       // class not className!